Add search function to header
